📦 chore(Dockerfile): move hashed-parachain binary to /usr/local/bin 📝 docs(Dockerfile): update CMD to use hashed-parachain binary 📝 docs(config.toml): update default_command and command to use hashed-parachain binary 📝 docs(k8-manifests/zombienet-local.toml): update default_image and image to use correct Docker images
The Dockerfile has been modified to move the hashed-parachain binary to the /usr/local/bin directory. This ensures that the binary is accessible from anywhere within the Docker container. The CMD instruction in the Dockerfile has been updated to use the hashed-parachain binary as the entry point for the container.

The config.toml file has been updated to use the correct command for the relaychain and parachains. The default_command and command now use the hashed-parachain binary.

The zombienet-local.toml file has been updated